Update hasChangesIn to fix PR Jenkinsfile
This commit updates the hasChangesIn function in utils.groovy to fix
Jenkinsfile.pull-request (it also moves Jenkinsfile.pull-request to the
jenkins folder). The --quiet-only option should not be used when runnin
git diff, since we won't be able to grep for the given String. Also,
update the check on grep's return status to 0 since 0 should be returned
if lines are found.