return total_cols when retrieiving col data
instead of passing the total number of cols when the data item is first created, pass the total number of columns each time we retrieve column data.

The total number of columns can change when the object associated with a data frame is reassigned, so it should be passed along with column data.