fix data viewer column navigation bug (#13270)
* refactor maxDisplayColumns and maxColumns logic

* WIP data viewer bugs

* fix broken indexing and column count

* add element id to data viewer column number input

for test automation!

* add tests + extra handling for `describeColSlice`

* cleanup debug prints

* return total_cols when retrieiving col data

instead of passing the total number of cols when the data item is first created, pass the total number of columns each time we retrieve column data.

The total number of columns can change when the object associated with a data frame is reassigned, so it should be passed along with column data.

* clean up DataViewer.cpp debug logging

* clean up dtviewer.js logging

* more cleanup and commenting

* use `totalCols` consistently for searchability

instead of `totalColumns`

* use previous logic for `isLimitedColumnFrame`

* update NEWS.md

* set help preview `max_cols` based on user pref