Speedup and bugfix for rsession metadata
* Don't create the session directory as part of creating the storage object
as this implicitly creates a session that's not valid just by looking to see if
a session exists with a given id.

* Don't use ensureDirectory for the properties directory on each property access.
This adds overhead and improperly creates a session just by trying to read a
property from it. Instead, catch the "no such file or directory" error when trying
to write a property and create the directory at that point.

* Bonus - speedup of test for web-socket upgrade (don't create the regex on every
request) - this showed up in the flame-graph I did of the rserver under load.