move `detectRosetta` call to desktop header init
Moving the call later in the IDE startup ensures that i18n is already initialized, so that the Rosetta check dialogs can display in the correct language.

I've also removed an initialization call to i18n, which is not needed (and not recommended to call more than once in the i18next docs).