Reduce number of calls to canSuspend/allowSuspend
Split the checkForSuspend method into two pieces:
  1) run before processing each request to check for explicit suspension - it only
calls allowSuspend if the user requested suspension
  2) another that's run to check for idle suspension that's run periodically.
The rsession waitForMethod, poll handler has a 50ms timeout for background processing
while waiting for r console input.  Only if 20 of these go by now do we check for
an idle session (i.e. 1/second).

Are there scenario where console input is being processed and we need
allow suspension?  Could use a timer here but trying to avoid
extra system calls since this is an inner loop for running R code.

Or we could a check in the offline polling thread so it's done
regularly, not based on the rate of console input.