Addresses: https://github.com/rstudio/rstudio-pro/issues/4922 (#13647)
In Posit Workbench, updating the 'executing' metadata property
too frequently causes performance issues, especially with some
customers that have network file systems that limit the number of
writes to a file per second. In this PR the rate of writes it not
controlled by the number of console-input statements but instead
limited to a few times a second. The transition to the executing
state happens right away, but the transition from executing to idle
happens in the offline service at no more than 4X second, so no
more than 8X/second.