Allow duplicate connection pane entries from pro drivers
Consider ODBC drivers with the Installer of "RStudio Pro Drivers" or
"Posit Pro Drivers" the same as those with an installer of "RStudio".
The "RStudio" installer indicates that the driver was installed by
Desktop Pro.

Previous to this change, a driver from the pro drivers would not be
listed in the list of connections if its name was identical to another
driver.