Set path of desktop driver install to match original name
This change will ensure that drivers installed by RStudio Desktop are
installed in the path that matches the name of the driver as supplied by
the installation file.

In cases where the deduplication logic appended the `with Posit Drivers`
the directory created would include this suffix and be space delimited.
We want to keep the original name for the directory, which has to be
processed during the installation, finding, and uninstallation of the
driver.

A new helper function .rs.connectionStripRStudioDriver is added to
complement the existing .rs.connectionOdbcRStudioDriver