Find appropriate ODBC Driver settings ini file
This supports installing and configuring the Databricks driver, which
under the hood is the Spark ODBC driver.

The Spark driver will only read the `rstudio.sparkodbc.ini` file, but
the IDE installation code was modifying the `rstudio.databricksodbc.ini`
file.

This allows us to stop bundling the additional ini file that was added
for the original workaround.