Fixed intermittent bug with auto-save dialog error
This happened in the load/stress tests when there was a
race between one request updating the properties file and
another trying to read it. Once in a while it would see it
as empty and report that auto-save error to the user.
We put in a quick retry in that case and it resolves the problem.

Reviewed in PR: https://github.com/rstudio/rstudio-pro/pull/6468