Increasing these parameters to improve scalability for Rstudio server
Addresses: https://github.com/rstudio/rstudio-pro/issues/6149

- If you are using the server version with only one or two
users you might want to set www-thread-pool-size back to 2.
Using 6 threads as in this PR is slightly less efficient in terms of CPU cache, adds
some memory and context switching overhead, thus
a bit of latency for every request compared with only 2, but
the request thread blocks often enough that 2 is not enough to prevent
bottlenecks when multiple users are on the system.

- The rsession-proxy-max-wait-secs value may increase the time delay
for some errors, or increase latency in those occasional failover situations
but otherwise should avoid errors for everyone.