Single-thread proxy connections for openssl
Addresses: https://github.com/rstudio/rstudio-pro/issues/7264

Need to ensure the client to rserver AsyncConnectionImpl uses
the same strand as the AsyncClient used for the rserver to
backend for websocket connections that are "full duplex", i.e.
emit traffic at the same time on both sides. When using openssl,
boost::asio requires that handlers run on a single thread. Previously
AsyncClient itself got a strand to fix part of the problem but
this fixes the other side.

I did not start out using a pointer for pStrand but ended up
there because:

1) The strand class is not copyable and requires an io_service as
a constructor arg that's also a constructor param.
2) The AsyncClient is extended by several other classes with default params
3) Most callers don't need the strand
4) We don't need the strand to be in place until the 'execute'
5) The AsyncConnection and AsyncClient are destroyed at the same time so the pointer
should not stick around after it is stale.