Backport mac ami changes to rel-cranberry-hibiscus (#15602)
* Use Mac AMI for builds (#15575)

Jenkinsfile changes to target the EC2 Mac AMI instead of macincloud, and changes to dependency scripts and cmake files to stop installing homebrew under the home directory (we now install it in the standard locations as part of the machine image).

* skip test on Mac (#15584)

This test is failing on our Mac builder. Might be a timing issue, it was working on a faster build agent, but failing on the slower one.

* don't try to create keychain if it already exists (#15591)

* don't try to create keychain if it already exists

* tweak message

* retry checkout 5 times with a pause before each try (#15597)

* Clean keychain before pulling (MacOS jenkinsfile) (#15598)

* clean keychain before pulling repo
* remove stray line