Redo how memory limits are managed in RStudio
This address the pro issue: https://github.com/rstudio/rstudio-pro/issues/5019
where quarto 1.6 is incompatible with our virtual memory limit logic.

- Adds a UI for warning the user, providing an error dialog, and
(optionally) aborting the session when it's low on memory

- Changes core library setResourceLimit(MemoryLimit,..) from using virtual
memory (ulimit -v) to Resident memory (ulimit -m). Virtual memory is inaccurate
but resident memory is only enforceable with more modern linux kernels that
support cgroups, and extra logic to manage the groups.

- Enforces memory limit in RStudio itself, by using the existing memory usage
widget polling, adding a few fields to the memory changed event.

- Added options to rsession.conf so an admin can turn on/off the 'abort' session
feature and adjust the limit of system memory that triggers an abort. This is
especially useful for testing, so you can trigger that dialog without getting too close
to the limit that breaks the test.

- Currently mac/windows return 0 for the process limit but if it would be useful we could
return OS specific limits as appropriate.