Update open file maximum with child processes (#16143)
* Switch to using the soft limit when closing file descriptors in a child process.
* Establish a reasonable upper limit of maximum files when closing file descriptors in a child process.

In recent Linux/Docker systems the hard limit for open files is set to nearly 1 billion. This causes issues with our slow close fallback mechanism which tries to brute-force close every file descriptor up to that limit. The limit of 1M included in this change was taken from an Ubuntu 22.04 system.