Avoiding set_default_verify_paths with a custom CA (#16762)
In a URL Mallet test I ran that mainly proxied requests to sessions,
this api was using > 80% of the CPU, much of it in spin-locks so it looks like
it locks around the code to read and parse all of the default CAs.

Since our session ssl uses custom CAs anyway, we can just skip that.
On my linux system with 32 cores, I was seeing 600% CPU with a 10 thread
URL mallet test before the fix. Afterwards, it never went above 150% and
the test finished faster.