Added a cache around ssl contexts
- this will significantly speed up https client requests made by
the rsession. Currently, for each request we search the default CA
registry and load up all of the certs.

In general, there will only be one ssl context that's cached. For any
downstream dependencies that use lots of custom CAs, you'll need to use
removeCertFromCache when that custom CA is no longer needed (or periodically
use the api to clear the cache).